pyproject.toml
[project]
name = "pytris"
version = "0.1.0"
description = "A Clone-tris in python"
readme = "README.md"
authors = [
    { name = "Abderrahmane Tahri Jouti", email = "tj.abderrahmane@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.5",
]

[project.scripts]
pytris = "pytris:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.basedpyright]
reportUnusedCallResult = "none"
